+++ mulle-bootstrap-tag.sh
@@ -45,78 +45,58 @@ name=`basename -- "${PWD}"`
 usage()
 {
    cat <<EOF
-usage: tag [tag] [vendortag] [vendorprefix]
+usage: tag [-f] <tag>
+
+   -d           : delete tag
+   -f           : force tag
 
    tag          : the tag for your repository ($name)
-   vendortag    : the tag used for tagging the fetched repositories
-   vendorprefix : prefix for the vendortag, if vendortag is "-"
 EOF
 }
 
 
-check_and_usage_and_help()
-{
-   if [ "$TAG" = "" -o "$TAG" = "-h" -o "$TAG" = "--help" -o "$VENDOR_TAG" = "" ]
+GIT_FLAGS=
+TAG_OPERATION="tag"
+
+
+while :
+do
+   if [ "$1" = "-h" -o "$1" = "--help" ]
    then
       usage >&2
       exit 1
    fi
-}
 
-
-project=`find_xcodeproj "${name}"`
-AGVTAG=
-
-if [ "${project}" != "" ]
-then
-   log_fluff "Trying agvtool to figure out current version"
-   dir=`dirname -- "${project}"`
-   [ -x "${dir}" ] || fail "${dir} is not accesible"
-
-   AGVTAG=`(cd "${dir}" ; agvtool what-version -terse ) 2> /dev/null`
-   if [ $? -ne 0 ]
+   if [ "$1" = "-f" ]
    then
-      log_fluff "agvtool failed"
-      AGVTAG=
-   else
-      log_info "Current version: ${AGVTAG}"
+      GIT_FLAGS="${GIT_FLAGS} ${1}"
+      TAG_OPERATION="force tag"
+      [ $# -eq 0 ] || shift
+      continue
    fi
+
+   if [ "$1" = "-d" ]
+   then
+      GIT_FLAGS="${GIT_FLAGS} ${1}"
+      TAG_OPERATION="delete the tag of"
+      [ $# -eq 0 ] || shift
+      continue
+   fi
+
+   break
+done
+
+
+TAG=${1}
+[ $# -eq 0 ] || shift
+
+
+if [ -z "${TAG}" ]
+then
+   usage >&2
+   exit 1
 fi
 
-
-TAG=${1:-"$AGVTAG"}
-[ $# -eq 0 ] || shift
-
-VENDOR_TAG="$1"
-[ $# -eq 0 ] || shift
-
-VENDOR_PREFIX="$1"
-[ $# -eq 0 ] || shift
-
-
-if [ -z "${VENDOR_TAG}" -o "${VENDOR_TAG}" = "-" ]
-then
-   if [ -z "${VENDOR_PREFIX}" ]
-   then
-      VENDOR_PREFIX=`basename -- "${PWD}"`
-      VENDOR_PREFIX="${VENDOR_PREFIX%%.*}"  # remove vile extension :)
-   fi
-
-   if [ "${VENDOR_PREFIX}" = "-" ]
-   then
-      VENDOR_TAG="${TAG}"
-   else
-      VENDOR_TAG="${VENDOR_PREFIX}-${TAG}"
-   fi
-
-   check_and_usage_and_help
-
-   log_info "Will use \"${VENDOR_TAG}\" to tag clones"
-else
-   check_and_usage_and_help
-fi
-
-REPO="."
 
 
 git_tag_unknown()
@@ -124,7 +104,7 @@ git_tag_unknown()
    local name
    local tag
 
-   name="${1:-${PWD}}"
+   name="${1}"
    tag="${2}"
 
    if [ ! -d .git ]
@@ -144,14 +124,15 @@ git_tag_unknown()
 git_must_be_clean()
 {
    local name
-   local clean
 
-   name="${1:-${PWD}}"
+   name="$1"
 
    if [ ! -d .git ]
    then
       fail "\"${name}\" is not a git repository"
    fi
+
+   local clean
 
    clean=`git status -s`
    if [ "${clean}" != "" ]
@@ -168,15 +149,17 @@ git_must_be_clean()
 
 ensure_repos_clean()
 {
+   local clonesdir
+
+   clonesdir="$1"
+
    #
    # Make sure that tagging is OK
    # all git repos must be clean
    #
-   (cd "${REPO}" ; git_must_be_clean "${REPO}" ) || exit 1
-
-   if  dir_has_files "${CLONES_SUBDIR}"
+   if dir_has_files "${clonesdir}"
    then
-      for i in "${CLONES_SUBDIR}"/*
+      for i in "${clonesdir}"/*
       do
          # only tag what looks like a git repo
          if [ -d "${i}/.git" -o -d "${i}/refs" ]
@@ -191,64 +174,55 @@ ensure_repos_clean()
 ensure_tags_unknown()
 {
    local tag
-   local vendortag
+   local clonesdir
 
-   tag="$1"
-   vendortag="$2"
+   clonesdir="$1"
+   tag="$2"
 
    #
    # Make sure that tagging is OK
    # all git repos must be clean
    #
-   (cd "${REPO}" ; git_tag_unknown "${REPO}" "${tag}" ) || exit 1
-
-   if  dir_has_files "${CLONES_SUBDIR}"
+   if dir_has_files "${clonesdir}"
    then
-      for i in "${CLONES_SUBDIR}"/*
+      for i in "${clonesdir}"/*
       do
          # only tag what looks like a git repo
          if [ -d "${i}/.git" -o -d "${i}/refs" ]
          then
-            (cd "${i}" ; git_tag_unknown "${i}" "${vendortag}" ) || exit 1
+            (cd "${i}" ; git_tag_unknown "${i}" "${tag}" ) || exit 1
          fi
       done
    fi
 }
 
 
-
 tag()
 {
+   local clonesdir
+   local tag
+
+   clonesdir="$1"
+   [ $# -eq 0 ] || shift
+   tag="$1"
+   [ $# -eq 0 ] || shift
+
    local i
-   local script
 
-   run_fetch_settings_script "pre-tag"
-
-   script=`find_fetch_setting_file "bin/tag.sh"`
-   if [ -x "$script" ]
+   if dir_has_files "${clonesdir}"
    then
-      run_script "$script" "${TAG}" "${REPO}" || exit 1
-   else
-      log_info "Tagging \"`basename -- "${REPO}"`\" with \"${TAG}\""
-      ( cd "${REPO}" ; exekutor git tag "${TAG}" ) || exit 1
-
-      if  dir_has_files "${CLONES_SUBDIR}"
-      then
-         for i in "${CLONES_SUBDIR}"/*
-         do
-            if [ -d "$i" ]
+      for i in "${clonesdir}"/*
+      do
+         if [ -d "$i" ]
+         then
+            if [ -d "${i}/.git" -o -d "${i}/refs" ]
             then
-               if [ -d "${i}/.git" -o -d "${i}/refs" ]
-               then
-                  log_info "Tagging \"`basename -- "${i}"`\" with \"${VENDOR_TAG}\""
-                  (cd "$i" ; exekutor git tag "${VENDOR_TAG}" ) || fail "tag failed"
-               fi
+               log_info "Tagging \"`basename -- "${i}"`\" with \"${tag}\""
+               (cd "$i" ; exekutor git tag $GIT_FLAGS "$@" "${tag}" ) || fail "tag failed"
             fi
-         done
-      fi
+         fi
+      done
    fi
-
-   run_fetch_settings_script "pre-tag"
 }
 
 
@@ -257,20 +231,25 @@ main()
    log_fluff "::: tag :::"
 
    ensure_consistency
-   ensure_tags_unknown "${TAG}" "${VENDOR_TAG}"
-   ensure_repos_clean
-
-   echo "Will tag `basename -- "${PWD}"` with ${TAG}" >&2
-   if  dir_has_files "${CLONES_SUBDIR}"
+   if [ -z "${GIT_FLAGS}" ]
    then
-      echo "Will tag clones with ${VENDOR_TAG}" >&2
+      ensure_tags_unknown "${CLONES_SUBDIR}" "${TAG}"
+   fi
+   ensure_repos_clean "${CLONES_SUBDIR}"
+
+   if dir_has_files "${CLONES_SUBDIR}"
+   then
+      echo "Will ${TAG_OPERATION} clones with ${TAG}" >&2
+   else
+      log_info "There is nothing to tag."
+      return 0
    fi
 
-   user_say_yes "Is this OK ?"
-   if [ $? -eq 0 ]
-   then
-      tag
-   fi
+   run_fetch_settings_script "pre-tag"
+
+   tag "${CLONES_SUBDIR}" "${REPO}" "${TAG}" "${VENDOR_TAG}" "$@"
+
+   run_fetch_settings_script "pre-tag"
 }
 
 main "$@"